# src/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for interacting with lighting hub simulator for use in the Josh.ai Backend Engineer Coding Challenge.\nNote: Try it Out will only work if the server is running on localhost:8080.",
        "title": "Josh.ai Coding Challenge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/lights": {
            "get": {
                "description": "Get list of all lights in the system.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lights"
                ],
                "summary": "Get summary of all lights.",
                "operationId": "get-lights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.LightConcise"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new light to the system.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lights"
                ],
                "summary": "Add a light.",
                "operationId": "add-light",
                "parameters": [
                    {
                        "description": "Full state of light to add",
                        "name": "light",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Light"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Light"
                        }
                    },
                    "400": {
                        "description": "invalid light data in body",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResp"
                        }
                    }
                }
            }
        },
        "/lights/{lightID}": {
            "get": {
                "description": "Get detailed state of a light in the system.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lights"
                ],
                "summary": "Get details about a light.",
                "operationId": "get-light-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of light",
                        "name": "lightID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Light"
                        }
                    },
                    "404": {
                        "description": "light not found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResp"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the state of a light in the system by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lights"
                ],
                "summary": "Update a light.",
                "operationId": "update-light-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of light",
                        "name": "lightID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "State fields to update",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LightUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Light"
                        }
                    },
                    "400": {
                        "description": "invalid fields in body",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "light not found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a light from the system by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lights"
                ],
                "summary": "Delete a light.",
                "operationId": "delete-light-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of light",
                        "name": "lightID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "light not found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.Light": {
            "type": "object",
            "properties": {
                "brightness": {
                    "description": "Light brightness 0-255",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "on": {
                    "type": "boolean"
                },
                "room": {
                    "type": "string"
                }
            }
        },
        "main.LightConcise": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                }
            }
        },
        "main.LightUpdate": {
            "type": "object",
            "properties": {
                "brightness": {
                    "description": "Include to set light brightness 0-255",
                    "type": "integer"
                },
                "name": {
                    "description": "Include to set light name",
                    "type": "string"
                },
                "on": {
                    "description": "Include to set light on/off",
                    "type": "boolean"
                },
                "room": {
                    "description": "Include to set light room",
                    "type": "string"
                }
            }
        }
    }
}